ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
sqlalchemy==2.0.23
httpx==0.26.0
